DYLD/LD vars cannot be path-updated on OSX if SIP-enabled
Path-variable handling is not effective on OSX for DYLD and LD variables
in case System Integrity Protection (SIP) id enabled: in this situation
these variables are not exported in subshell context, so they appear
undefined to modulecmd.tcl.

Nullify modshare variable content for the moment, do not print warning
message. Adapt the DYLD/LD test cases when running on OSX if SIP is
detected enabled.