Detect inconsistent loaded environment state
When needing to correlate the information from the LOADEDMODULES and
the _LMFILES_ environment variables.

Does not affect module commands not requiring correlation of these
informations, like whatis/avail. Error raised on load, unload, switch,
reload, purge, list, save and restore commands.

May affect info-loaded or is-loaded commands if module passed as
argument to these command is specified as a full path modulefile.