Remove unneeded symlink transformation for Cygwin
Symlink created on Cygwin does not get an automatic ".lnk" suffix
extension. ".lnk" shortcut made on Windows cannot be interpreted as a
symlink either on Cygwin or Windows cmd. Tcl will not recognize a
directory or a modulefile when looking at these ".lnk" shortcut.

Remove the .lnk file name transformation as it is not useful. In
addition, 'isWin' returns false on a Cygwin environment as detected
platform on this tool equals to 'unix' not 'windows'.