Fix module func parameter expansion on CMake
To correctly handle parameter containing whitespace character or empty
elements. For the latter case, command definition must be adapted to the
number of args to be able to pass, to some extend (<5 args) empty string
element to modulecmd. There is no other way to achieve empty string
element argument as empty strings are skipped from ${ARGV}. [1]

[1] https://gitlab.kitware.com/cmake/cmake/issues/16341