Cache findModules results
Introduce a memory cache for the results build up by findModules
procedure. Done in order to reuse these results if next searches are
compatible instead of re-walking the filesystem.

A cached result may be used by a search looking for a subset of this
result. For instance searching for modules 'foo' in '/path/to/dir' may
use the cache result saved after the search of '*' modules in
'path/to/dir'. This superset result is then filtered-out whithin the
getModules procedure.