Unset any ref cnt var on setenv/unsetenv in any mode
Whatever the evaluation mode, unset any existing reference counter
variable when running setenv/unsetenv commands.

Adapt and add tests to put in light behaviors when unloading a
standardly-set variable going through append-path/prepend-path. The
standardly-set variable is transformed into a path-like variable through
the evaluation, resulting in the end in the definition of a reference
counter variable (if variable is not setenv/unsetenv during evaluation)
or in the deletion of this reference counter variable (if
setenv/unsetenv during evaluation).