Adapt interp alias commands to the evaluation mode
Make interpreter alias command targets vary depending on the evaluation
mode to adapt behavior of modulefile commands to the currently run
module command. For instance a 'setenv' command is evaluation as an
'unsetenv' command when unloading a modulefile.

This commit lays the groundwork to adapt modulefile commands to the
evaluation mode. All commands are kept for the moment in the situation
where the same procedure is applied to all evaluation modes.