Move 'system' to the per-mode evaluation interp
'system' modulefile command targets system procedure on load and unload
modes of modulefile evaluation. reportCmd is targeted on display mode.
For other modes a nop is targeted.

'system' can also currently be called from a modulerc evaluation when
load or unload mode is detected. Kept as is for the moment. May/should
change for next major release.