Trace exec to reportCmd on display mode enabled cmds
For those commands enables on display mode (commands that are also run
on this mode, not only reported, like */*-path commands): use an
execution trace (at leave step) to call for the report commands after
running the enabled procedure.

With this change these enabled-on-display commands do not have to handle
their display reporting anymore. Current evaluation mode can now be
ignored by these procedures.

The reportCmdTrace procedure is introduced to be set as the execution
trace procedure in case alias command for modulefile interp is not set
to reportCmd on a display evaluation mode.

As a side effect of this change, commands are reported even if their
execution raise an error (helps to see the command line along with
raised error message).