Set *setenv cmds as init-if-undef proc on whatis mode
Define a specific procedure for the setenv and unsetenv commands on a
whatis evaluation mode: setenv-wh.

This procedure checks the environment variable targeted by the command
and initialize it as an empty string if undefined.

Proceed this way to reduce the total amount of time of a whatis query
but still initializing variables if undefined to avoid 'undefined
variable' errors during evaluation if environment variable value are
referred to later in modulefile.

Value set to variable are not the correct one (no value set if variable
defined, empty string set if variable is undefined) but it should not
affect the whatis query as only the 'module-whatis' commands are
valuable in this mode.