Pass cur itrp ref in initInterpCommands to aliases needing it
Sets in initInterpCommands procedure interpreter reference as first
argument to interpreter alias requiring it, rather doing it in
initModfileModeAliases where actual interpreter reference is not known
(as this procedure is called at first time a mode-specific interp is
created, but not on subsequent times when another interp of same mode
but working in a nested level is build up).

Now initModfileModeAliases defines a specific __itrp__ keyword to let
initInterpCommands know that the argument to pass to the alias to define
is the actual interp reference.