Ignore '--notuasked' flag in case it is passed on module load
--notuasked flag enables to keep track of the user asked state in saved
collection (by flagging those not directly asked by user). Ignore it if
passed directly to a 'module load' command (do not raise error) in case
a collection file is sourced like a shell script.

Do not advertise this new flag as it has no effect and should be kept
internal as user asked mechanism is automatically controlled from the
location the module load command is issued.