Setenv on unload mode set var to val in interp context
Interpreting a setenv modulefile order during an unload evaluation,
variable is still set to be unset in generated shell code but it is set
to the value defined on the setenv order in the interpreter context
instead of being cleared.

This change is made to recover and improve behavior of compat version
and previous version of modulecmd.tcl where variable's value found in
environment were not touch in interpreater context.

This is useful when using environment variable value to
append/prepend-path another variable. This change enables to correctly
unset this other variable on an unload evaluation. Was previously not
unset since the first variable used to edit the second was found
cleared.

This change is currently limited to the setenv order and does not apply
to the prepend-path and append-path, where during an unload evaluation
variable value part is removed from variable also in interpreter
context. Need to distinguish value in interpreter context from value to
set in shell output code to also apply this mechanism to path edition
modulefile orders.