Build and maintain dep chain between loaded mod
Introduce setModuleDependency and unsetModuleDependency procedures to
build and maintain global arrays representing the dependency chain
between loaded modulefiles based on the prereqs defined by each of these
modules.

setModuleDependency is called when module is loaded or current
environment state is analyzed (called by setLoadedModule).
unsetModuleDependency is called when module is unloaded (called by
unsetLoadedModule).

The g_moduleDepend array is set for each loaded module as a mirror of
their relative definition in g_loadedModulePrereq, where module
reference in the prereq definition is replaced by the name of the actual
loaded modulefile matching the module reference.

When a module reference of a prereq definition has no match among the
currently loaded modulefiles, this unmet dependency is registered
instead in the g_moduleUnmetDep global array. This situation happens
when an optional dependency is not satisfied or if the load of a module
has been forced (no dependency loaded prior to the load of the asked
module).

Reverse arrays are also set (g_dependHash and g_unmetDepHash) to easy
the search of registered elements.

When a new module is loaded, setModuleDependency looks if this new
module was previously set as an unmet dependency. If it is the case,
update dependent loaded modules dependency definition now the unmet dep
is satisfied.