Rework code to get unload autoloaded mods when no more req
Add getRequiredLoadedModuleList proc to return all loaded modules in the
requirement list of passed mod. This proc is called once mod has been
unloaded, prior to the deletion of these requirement info in
cmdModuleUnload.

Add getUnloadableLoadedModuleList proc, to return the modules that have
been automatically loaded and could now be automatically unloaded, as no
more modules requires them (mandatory or optionaly). This proc takes the
result of getRequiredLoadedModuleList as input.