Rework is-loading to get closer to is-loaded handling
Rework is-loading procedure to call getLoadedMatchingName procedure like
is-loaded. Adapt getLoadedMatchingName to work on loading modules rather
loaded modules. To achieve that, isModulefileLoading,
getModuleFromLoadingModulefile, getLoadingModuleFileList and
getLoadingModuleList procedures are introduced to be called instead of
their 'loaded' counterpart when getLoadedMatchingName is called to look
at loading modules.

Remove is-loading procedure from aliases set for modulefile interpreter
as there is no current need to acces such kind of procedure during a
modulefile evaluation context.