Track loaded mod violating their prereq/conflict def
Update setModuleDependency and unsetModuleDependency procedures to
register/unregister in a ::g_prereqViolation array loaded modules
violating their prereq definition (means they have been loaded in force
mode or their prereq was unloaded in force mode).

Proceed in a similar way for conflicts with a ::g_conflictViolation
array, still in setModuleDependency and unsetModuleDependency procedures.

For prereq violation, updates of dependent loaded module eventual
violation is obtained by the re-compute of the dependency chain
performed in setModuleDependency and unsetModuleDependency.

For conflict violation, updates of violation state is done for loaded
modules conflicting with module processed by setModuleDependency or
unsetModuleDependency. To achieve that list of conflicting-with loaded
modules is set in ::g_conflictViolation array. This array is handled by
setModuleConflictViolation and unsetModuleConflictViolation procedures.

doesModuleConflict procedure is updated on the way to skip from its
result eventual own reflexive conflict, as this procedure is now called
once mod is effectively loaded and not anymore prior mod load (as when
it is called from cmdModuleLoad). Also check conflicting modules
declared by current module are loaded or not.