Update getDependentLoadedModuleList to return sane deps
Add a sat_constraint argument to getDependentLoadedModuleList procedure
to ask for dependent loaded modules whose constraints are satisfied.
This argument is enabled when reloading dependent modules in load,
unload and switch actions.

A areModuleConstraintsSatisfied procedure is introduced and
getDependentLoadedModuleList relies on it to determine if a given loaded
module has its constraints (prereq and conflict) satisfied as well as
all loaded modules whose given loaded module depends on.

When no module name is passed to areModuleConstraintsSatisfied, this
procedure returns if all loaded module satisfy their loading
constraints.