Introduce a 'force' arg to cmdModuleUnload
Add the ability to by-pass requiring modules check when unloading a
module with an argument on the cmdModuleUnload procedure.

This new 'force' argument is enabled when reloading modules as the
requirement by-pass will immediately be fixed once the required modules
are loaded again.

Happens when reloading dependent modules on load, unload and switch
actions. Some of the dependent modules may be excluded from reloading
module list as they are in a prereq/conflict violation state so the
remaining modules to reload may face some requiring modules still
loaded, thus their reload need to be forced.

Also enabled when purging loaded modules as all modules should be
unloaded even if some violations make unloading prereqs before the
module requiring them.