Add 'being_unload' arg to getDependentLoadedModuleList
Indicate to getDependentLoadedModuleList procedure that targeting module
is being unloaded. In this case the procedure will return loaded modules
conflicting with passed mod and the dependent modules of these
conflicting module.

These loaded modules are added to the dependent loaded module list since
passed mod is being unloaded, they can be reloaded now the conflict
violation is fixed. This reload garanties the modules previously in
conflict are now correctly loaded and in an expected loaded order.