Take disapearing prereq in account on getDependentLoadedModuleList
When 'being_unload' argument is set, which means passed mod is currently
being unloaded, exclude from result list the loaded modules requiring
passed mod. As this module is being unloaded, these dependend modules
are entering in a violation state, so they cannot be reloaded.

This is achieved by temporarily defining the dependent modules in prereq
violation state, for the time of the getDependentLoadedModuleList
procedure execution.