Handle default version sets on an hidden modulefile
When looking for the default version of a given modulefile, if this
default is set on an hidden modulefile, this hidden modulefile were not
found as the initial modulefile search were not looking for hidden
modulefiles.

When this situation is now spotted, modulefile search is started again
with inclusion of hidden files.

Add a test for this situation in 20-locate/067-hidden.

Fixes #177