install: use sed rather grep and cut
Use sed command rather a combination of grep and cut in 'configure' and
'Makefile' scripts. As a result grep and cut commands are not required
anymore to install new Modules releases.

However, grep is still required to install compatibility version.

Big thanks to Michael Sternberg (@mgsternberg) for crafting the right
sed commands to achieve a more efficient installation process.

Fixes #175