Pass a module list to getDependentLoadedModuleList
Adapt getDependentLoadedModuleList procedure to get a module list as
argument rather a single module name.

Helps to get in one call all the dependent modules of those passed as
argument.

Also improve correctness as sometimes multiple modules may violate
constraint of another one. To get this module freed, all conflicting
modules may set unloaded. So if all of these conflicting modules are
passed as argument to getDependentLoadedModuleList, the dependent module
will be seen freed and will be part of the procedure result list.