Rely on cmdModuleUnload DU/URU/DR during switch
Rely on the DU/URU mechanisms of cmdModuleUnload during a switch action
and also the unload part of the DR mechanism.

cmdModuleUnload transmit the DR module list to cmdModuleSwitch for this
procedure to be able to handle the reload phase of the DR mechanism once
the newly switched-to module is loaded.

This change makes useless the being_reload argument of the
getDependentLoadedModuleList procedure, as module candidates are now
found when looking for DU modules in cmdModuleSwitch. being_reload
argument is thus removed from getDependentLoadedModuleList proc.