No prereq violation warn msg for DR mech or purge
During a purge or the unload phase of the DR mechanism do not warn about
the unload process violating a prereq constraint.

In these situation, violation is due to the dependent module not being
able to also unload or to unload prior its requirement. All of this
because this dependent module is its in a violation state.