No prereq violation set for dependent on getDependentLoadedModuleList
Modules strongly requiring mods where previously set in prereq violation
state as mod where unloaded. Remove that since these modules requiring
mod during an unload (the DU modules) are now part of the list of
modules passed as argument to getDependentLoadedModuleList. So this part
of the code will not be run anymore.