Add 'warn' arg to cmdModuleUnload proc
Add a 'warn' argument to the cmdModuleUnload procedure to control
whether or not display the warning message when a module cannot be
unloaded because it is still required by other loaded module.

This argument is equal to true every time except for the interpretation
of 'module load' commands in modulefile when mode is 'unload'. In this
situation module unload may not be performed if module is still required
but it is a normal situation like the UReqUn mechanism (only the useless
modulefiles are automatically unloaded).