Auto handling and 'module load' commands in modulefile
'module load' commands in modulefiles may impact the auto handling mode,
so a few things need to be adapted:

- on a 'load' mode, register each module passed to the module load
command as a prereq of evaluated modulefile. This way the dependency
will correctly be tracked

- on an 'unload' mode, proceed to the unload action when evaluating
the 'module load' statement only if auto handling is disabled. Do
nothing when evaluating the statement if auto handling is enabled, as
these modules will be unloaded as part of the UReqUn mechanism. This way
all unload/reload are triggered from the top unload action to ensure
nothing will get in the way of the DepUn/DepRe/UReqUn processed from
there.

Note that on an unload mode, modules are unloaded only if no other
loaded modules require them and they have not been asked by user,
elsewhere no unload is made to ensure the consistency of the loaded
environment (as these modules are explicitely required by user or by
another loaded module)

Treating these modules to unload through UReqUn mechanism leads to some
behavior changes:

- unload interpretation is not cascaded anymore in this case and every
module to unload are processed sequentially
- a consequence of the previous point is that an error happening when
processing the unload of a UReqUn module, will not unwinds the whole
unload process as it does when auto_handling is disabled
- unload happens in the reverse loaded order, not in the order defined
by the 'module load' commands in the modulefile
- as a consequence of this previous point, environment change after the
whole unload may be different if the unloaded modules compete over a
variable to set when unloading (last module to be unloaded will not be
the same)