ts: adapt to modfile 'module load' auto/no-auto handling
More test case adaptation to follow the new 'module load' commands
handling when the auto_handling mechanisms are enabled or disabled.

The tests changed here show different behavior whether auto_handling is
enabled or not. Due to the fact that module unloads is not cascaded when
auto_handling is enabled and UReqUn modules ares unloaded in reverse
loaded order, not in the order specified within modulefile. These 2
behavior changes lead to different result: command order change, unwind
process when a module to unload is erroneous, or variable resulting
value when unloading modules compete over same variable to set.