From warn/nuasked to onlyureq/onlyndep args on cmdModuleUnload
Change warn and nuasked arguments previously added to cmdModuleUnload
procedure by onlyureq and onlyndep to better cope needs.

'onlyureq' indicates only the not required anymore modules should be
unloaded (not user asked modules or those not required by another loaded
module). This flag is enabled when dealing with 'module load' commands
in modulefiles interpreted in unload mode. If a module cannot be
unloaded because it is required, then no error message is produced.

So 'onlyureq' argument replaces both 'warn' and 'nuasked' arguments.

'onlyndep' indicates that only modules not required by other modules
should be unloaded. It is used when dealing with 'module unload'
commands in modulefiles interpreted in load mode. This is useful when
force mode is enabled to avoid breaking the dependency of a loaded
module when trying to load a module declaring this dependency as a
module to unload (acts like for conflicts).