Only look at loaded modules when unloading
Adapt getPathToModule procedure to only look at the loaded module list
when searching a module to unload. If passed module not found in this
situation does not produce an "Unable to locate modulefile" issue.

So unloading an unexistent modulefile does not produce an error anymore,
as it was the case on compatibility version.

Moreover switching from an unexistent modulefile to another modulefile
does not produce the "Unable to locate modulefile" error and new module
load is attempted. Compatibility version does not walk in this situation
as on this version module to unload must be loaded elsewhere an error is
triggered.

Adapt testsuite and diff document.

Fixes #199