Record messages to report them by block on load/unload
Save warning, error and internal bug message produced during a
cmdModuleLoad or cmdModuleUnload evaluation for a module to report all
these messages as a dedicated "Loading <module>" or "Unloading <module>"
block after evaluation.

A recording ID is set when preparing a load or an unload evaluation. If
a recording ID is found when running reportError, the message is
recorded in a global structure rather reported immediately. Once
evaluation is finished, all messages are rendered under a block.

Improve this way readability of those messages when evaluating mode may
cascade a lot of stuff.