Compute loaded mod dependency relations without cycle
When computing dependency relations between loaded modules and a freshly
loaded one, do not launch a refresh of those unmet relations for an
already loaded module with the new one if this relation closes a
dependency cycle.

When parsing the already loaded environment, the loaded module
dependency relations are computed giving each module only the visibility
of modules loaded prior it. Modules loaded after one module are not
taken into account as they have not taken part of its requirements.