Rework ref counter handling and relative error mngt
No more disturbing warning messages when reference counter variable is
not coherent with its relative environment variable. Erroneous value
will raise a debug message only and relative reference counter is
ignored (which means a count of 1 is applied).

Now environment variable is trusted and if its path elements are found
in a reference counter variable, this counter is taken into account.
Elsewhere, a reference counter of 1 is applied.

Which means a value found in reference counter variable but not in
variable is ignored. A value found in variable but not in reference
counter (or with a bad reference counter value) gets a count of 1
applied.

Adapt existing tests in 70/152 and 50/40 and add some unload command
tests in 50/040.

Fixes #206