Call readModuleContent in checkValidModule proc
To read a modulefile so the code to read file and analyze its header to
check it is a valid modulefile is just written once.

readModuleContent procedure has been adapted to proceed like
checkValidModule did when it is called in a only_check_validity mode. In
this mode, only the first line of the file is read as we just want to
parse the header of the file (as it was done previously in
checkValidModule procedure).