Introduce getFilesInDirectory proc to get dir elt list
Extract code from findModules procedure that analyzes and returns the
content of a given directory to define a dedicate procedure for that
job: getFilesInDirectory.

This getFilesInDirectory procedure will also be made available as a
command defined in the Tcl extension library to reduce the IO calls
made to achieve the task.