Introduce --indepth/--no-indepth avail mode
Add the ability to perform non-recursive avail search by limiting
results to the elements matching search query and not including elements
contained in the directories part of the result.

Default behavior is unchanged: in depth search is performed. --indepth
and --no-indepth command-line switches are introduced to control
behavior as well as an environment variable: MODULES_AVAIL_INDEPTH.
Default behavior may also be controlled at ./configure time with the
--enable-avail-indepth and --disable-avail-indepth switches.

The --no-indepth mode is pretty useful to provide an efficient glance
at the available packages when a lot of different versions exist for
each of them. Shell completion scripts also benefit from the
non-recursive view to propose choices corresponding to the current depth
level only. Moreover I/O operations are reduced on this --no-indepth
mode as the full modulefile tree is not entirely walked through.

This commit introduces the new configure script and modulecmd options and
it lays the ground for the changes on the module find and the result
display that will be made in the next commits.

Fixes #150.