Adapt getModules/listModules to --no-indepth mode
Remove in getModules the result entries collected within the procedure
where more filename path separator are found than in search query
string. This way entries related to the content of directories matching
query are removed. Only entries matching search query at its very same
depth level are kept + the matching directories holding those entries.

This or these matching directories holding entries matching search query
are kept to get access to the default or the latest entries in these
directories if -d (default) or -L (latest) filters are enabled.

The directories holding entries are then excluded from display output in
listModules procedure to keep this --no-indepth mode displaying only
entries lying at the same directory depth level.

When the -d (default) or -L (latest) filters are enabled and the
--no-indepth mode is enabled too, if search query targets the root
content of the modulepath, do not apply the output filter. In this
situation every elements at the root of a modulepath can be considered a
default of a latest entry.