Optimize findModules when no in depth search is asked
Cut down some I/O operations in findModules procedure when a no in depth
search is performed as it is not required to check in depth existing
modulefiles, only those matching query depth level.

.modulerc files are still searched in depth, as they can produce virtual
modules and in the future they may invalidate modulefiles. All files at
the search query depth level are analyze. And for the directories below
this depth level modulefiles are analyzed until one is found valid
(other modulefiles are skipped as we only want to determine if directory
holds at least one valid modulefile and should be reported).