Introduce colored output
Add the ability to graphically enhance some part of the produced output
to improve result readability.

Colored output enablement could be controlled at configure time with the
--enable-color option, or with the MODULES_COLOR environment variable or
with the command-line switch --color. MODULES_COLOR variable and --color
switch accepts the following values: never, auto and always. With the
--enable-color configure option, the 'auto' color mode is applied. When
color mode is set to 'auto', output is colored if a terminal is attached
to stderr.

Colors to apply to each elements can be controlled at configure time
with the --with-colors option or with the MODULES_COLORS environment
variable. A color set consists of a list of key=value separated by
colon where each key represents an output element that can be colored
and each value represents a SGR code applied to this element to color
it.

Thanks to [1] for the detailled description of the Select Graphic
Rendition (SGR) codes.

[1] https://stackoverflow.com/questions/4842424/list-of-ansi-color-escape-sequences