Pass elts length to displayElementList proc
Externalize from the displayElementList procedure the computation of the
length of each element to display. This is done to prepare the output of
graphically enhanced elements from this procedure. Element lengths are
computed in each procedure building the display list. This is where the
graphical enhancement will take place so it is the right spot to learn
the real length (without SGR code) of these elements to output.

Take this opportunity to also change displayElementList procedure to pass
it the display list as a list argument rather as "args" which requires to
'eval' the procedure call.