Define a light and dark background default color set
Define by default a color set for dark background and a color set for
light background. These color sets can be define at configure time with
the --with-dark-background-colors and --with-dark-background-colors
options respectively.

Terminal background color kind can be defined to 'light' or 'dark'. This
value can be set at configure time with the --with-terminal-background
option or with the environment variable MODULES_TERM_BACKGROUND.

If the default color set is not superseded by the MODULES_COLORS
environment variable, the dark background colors will apply if terminal
background is set to dark. If the terminal background is set to light,
the light background colors will apply instead.