Evaluate mod then check conflict violation
Give the chance to the 'module unload' statements in modulefile to
solve conflict that were registered prior the load attempt of this
modulefile by evaluating modulefile then check afterward if a conflict
violation is there rather checking if a conflict violation exists (and
shout error if there is) before evaluating modulefile.

So modulefile is evaluated even if a conflict is registered against him
prior its evaluation. If a conflict is still set after its evaluation
(and the evaluation of its requirements) everything is rolled back.

The 'module unload' commands could solve these already registered
conflicts. 'conflict' commands used in the loading modulefile will
however still break the load immediately if the conflict is seen
violated to maintain same behavior.

With this change, behavior of these 'module unload' modulefile commands
is restored to what it was able to do prior conflict consistency
changes, yet this consistency is maintained as everything is rolled back
in the end if conflict is still there.

doesModuleConflict procedure has also been changed to look at loading
modules in addition to loaded modules, since the conflict check triggered
by 'module unload' will not raise an error (error is now raised based on
the result of the doesModuleConflict call.

Fixes #216