Track deps defined by 'module switch' in modfile
When 'module switch' commands are found in modulefiles, track
switched-off modulefile as a conflict and switched-to modulefile as a
requirement to apply same behaviors than for 'module load' and 'module
unload' commands in modulefiles.

If 'switch' sub-command has only one argument, do not define any
conflict, only a requirement on the defined switched-to argument.

CAUTION: it is not recommended to use the `switch` sub-command in
modulefiles as this command is intended for the command-line for a 2in1
operation. Could be removed from the modulefile scope in a future
release. Use `module unload` and `module load` commands in modulefiles
instead.

Fixes #229