Revert 'module switch' in modfile when unloading modfile
No distinction were made previously for 'module switch' command used in
modulefile regarding modulefile evaluation mode. This was a mistake as
an unload evaluation of a modulefile should revert what has been set at
load time.

During an unload evaluation, a 'module switch' command is converted to a
'module unload' of the switched-on module, like for 'module load'
command. Nothing is done for switched-off module, like for 'module
unload' command.

CAUTION: it is not recommended to use the `switch` sub-command in
modulefiles as this command is intended for the command-line for a 2in1
operation. Could be removed from the modulefile scope in a future
release. Use `module unload` and `module load` commands in modulefiles
instead.

Fixes #226