Find alt names of mod coming from distant dirs
If an alias is set on a directory and this directory default element is
an alias toward a modulefile. When this modulefile is loaded, it holds
no reference to first alias.

For default elements in a directory that points to a modulefile, when
this modulefile is loaded, it should receive the eventual aliases set on
the distant directory as alternative names.

Fixes #231