Fix mod selection for 'module load/switch' in modfile when unloading
When unloading a loaded module that contains 'module load' or 'module
switch' commands in its modulefile, loaded required module should be
unloaded if it has been automatically loaded.

This commit fixes the selection of this module to unload automatically
when evaluating in unload mode the 'module load' or 'module switch'
commands used in modulefiles. If a loaded module matches the 'module
load' or 'module switch' statement but was loaded after the module
making these statements, it should not be selected as the module to
automatically unload.

Fixes #232