Correct unloadable mod search when reverting 'module load/switch'
Select unload candidate from loaded module list excluding modules loaded
after the current unloading module. So only module loaded prior
dependent module are taken into account.

Then pass this candidate, if any, to cmdModuleUnload and rely on this
procedure to unload module only if it is unloadable.

Previous commit was relying on the list of unloadable modules, obtained
from getUnloadableLoadedModuleList procedure, which sometimes led to
corner cases especially if dependency information were not set somehow
(no module would be unloaded in that case as
getUnloadableLoadedModuleList relies only on these dependency
information).

As the candidate is selected to be the closest one from the module
requiring it, it changes unloading result when multiple candidates are
found prior requiring module (previously firstly loaded module was
unloaded).