Rework cmdModuleUnload code with break/continue commands
To improve clarity of the cmdModuleUnload procedure replace large
if/then/else conditional blocks by break or continue abortion commands.
A catch statement is set around the evaluation parts of the procedure
and if a break happens, saved settings are restored (which means some
evaluation went wrong and previous settings are rolled back).