Stop unload if one DepUn fails except if force mode set
During the unload of a module, if the unload of one of its dependent (by
the dependent unload mechanism) fails abort the whole unload process.
Exception made if the force mode is enabled. In this case failing module
stays loaded and Dependent Unload mechanism continues with next module to
unload.