If one UReqUn fails removes its requirements from UReqUn list
During the unload of a module, if the unload of one of its useless
requirements (by the useless requirement unload mechanism) fails,
removes the requirements of this failing module from the useless
requirement list so they are kept loaded as failing module will still be
loaded.

Such error is reported as a warning and it does not stop the whole
unload process as main module to unload and its dependent are able to
unload, just its requirements automatically loaded cannot be unloaded.
So the job has been done, not perfectly but main task is successful and
resulting user environment is coherent.

A getRequiredLoadedModuleList procedure is introduced to get all the
loaded modules required by a list of modules passed to the procedure.

Fixes #240