Stop if one DepRe unload fails except if force mode set
During the load or the unload of a module, if the unload of one of its
dependent (by the dependent reload mechanism) fails abort the whole
unload/load process. Exception made if the force mode is enabled. In
this case failing module stays loaded and Dependent Unload mechanism
continues with next module to unload. This failing module is removed
from the Dependent Reload list, so it will not take part of the load
phrase of the mechanism.

Fixes #239