Stop if one DepRe load fails except if force mode set
During the load or the unload of a module, if the load of one of its
dependent (by the dependent reload mechanism) fails abort the whole
unload/load process. Exception made if the force mode is enabled. In
this case failing module stays loaded and Dependent Unload mechanism
continues with next module to load.

When applying DepRe load phase during a switch command, consider force
mode a default since here we are trying to reload the dependent modules
found when unloading switched-off module. Switch action does not
guarantee to find all DepRe modules still loaded at the end of process.

Fixes #241