Register and report each evaluation and assoc context
Better track each module evaluation and the context associated to it in
order to report a more accurate information on the additional modules
loaded or unloaded when proceeding the main evaluation request.

When calling for a module load or unload, pass a context name to either
the cmdModuleLoad or cmdModuleUnload procedure called to let it register
the module evaluation and its associated context into a global structure
gathering all these informations per top-level evaluation.

Report messages on the additional evaluation occuring during a main
one are now based on these registered information rather on computed
list of elements passed to the cmdModuleLoad or cmdModuleUnload
procedures. As a result a more reliable information is obtained,
especially during a main load evaluation, where previously some
unload/load additional evaluation were not seen or were reported against
another automated mechanism.

Additional evaluation messages are only reported if the main evaluation
goes to its end, which means that if some event forces it to rollback
the summary list of actions will not be seen to let users better
understand the resulting situation (nothing was unloaded or loaded) and
focus on the error message. Previously finished auto mechanisms were
immediatly reported which confused people as an error occuring after the
mechanism leads to a full rollback.

In case an additional evaluation fails, the corresponding module is
removed from the list of modules on which the auto mechanism triggered a
load or an unload evaluation.

When unloading useless requirements in no-auto mode, these unloads are
now reported in their correct evaluation order (order of definition
found in modulefiles).

Fixes #244, fixes #245, fixes #246, fixes #247 and fixes #248