Report auto mech warn/err to top level message block
During an additional evaluation triggered by an automated module
handling mechanism, ensure warning and error messages are reported under
the message block of the main evaluation.

Add a 'recordtop' argument to the reportWarning and reportError
procedures in order to indicate that the message should be recorded
under the top-level message block instead of the currently active
message block.

Fixes #252.